refactor: update useClient import path to reflect new folder structure
